억까 캐치
원인은 안석x 이었씁니다 ^^
코드에는 문제 없구요
사람이 문제였습니다 ㅎㅎ
